split utils into file
